+++ crates/openshell-tui/src/ui/sandbox_settings.rs
@@ -4,7 +4,7 @@
 use ratatui::Frame;
 use ratatui::layout::{Constraint, Rect};
 use ratatui::text::{Line, Span};
-use ratatui::widgets::{Block, Borders, Cell, Clear, Padding, Paragraph, Row, Table};
+use ratatui::widgets::{Block, Borders, Cell, Padding, Row, Table};
 
 use super::draw_empty_message;
 use crate::app::{App, SandboxPolicyTab, SettingScope};
@@ -166,16 +166,7 @@ fn draw_confirm_set(frame: &mut Frame<'_>, app: &App, idx: usize, area: Rect) {
         ]),
     ];
 
-    let popup_height = u16::try_from(lines.len() + 2).unwrap_or(u16::MAX);
-    let popup = centered_rect(60, popup_height, area);
-    frame.render_widget(Clear, popup);
-
-    let block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(t.border_focused)
-        .padding(Padding::horizontal(1));
-
-    frame.render_widget(Paragraph::new(lines).block(block), popup);
+    super::draw_confirm_popup(frame, lines, t.border_focused, 60, area);
 }
 
 fn draw_confirm_delete(frame: &mut Frame<'_>, app: &App, idx: usize, area: Rect) {
@@ -204,16 +195,5 @@ fn draw_confirm_delete(frame: &mut Frame<'_>, app: &App, idx: usize, area: Rect)
         ]),
     ];
 
-    let popup_height = u16::try_from(lines.len() + 2).unwrap_or(u16::MAX);
-    let popup = centered_rect(55, popup_height, area);
-    frame.render_widget(Clear, popup);
-
-    let block = Block::default()
-        .borders(Borders::ALL)
-        .border_style(t.status_err)
-        .padding(Padding::horizontal(1));
-
-    frame.render_widget(Paragraph::new(lines).block(block), popup);
+    super::draw_confirm_popup(frame, lines, t.status_err, 55, area);
 }
-
-use super::centered_popup as centered_rect;
